ml-service/requirements.txt
tensorflow==2.15.0
keras==2.15.0
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
pytesseract==0.3.10
easyocr==1.7.0
pdf2image==1.16.3
opencv-python==4.8.1.78
requests==2.31.0
python-dotenv==1.0.0

